ChatSite/requirements.txt
django
channels-redis==2.4.2
channels==3.0.5
django-filter
djangorestframework
markdown
python-dotenv
Pillow